refactor: split worker.rs god-file into modules
