+++ homeboy.json
@@ -733,7 +733,7 @@
         "intra-method-duplication::src/core/runner/execution/failure.rs::IntraMethodDuplicate",
         "intra-method-duplication::src/core/runner/git_dependency_materialization.rs::IntraMethodDuplicate",
         "intra-method-duplication::src/core/runner/validation_dependencies.rs::IntraMethodDuplicate",
-        "intra-method-duplication::src/core/runner/worker.rs::IntraMethodDuplicate",
+        "intra-method-duplication::src/core/runner/worker/run.rs::IntraMethodDuplicate",
         "intra-method-duplication::src/core/upgrade/runners.rs::IntraMethodDuplicate",
         "near-duplication::src/core/artifact_contract.rs::NearDuplicate",
         "near-duplication::src/core/fuzz.rs::NearDuplicate",
@@ -803,7 +803,6 @@
         "structural::src/core/runner/lab/offload.rs::GodFile",
         "structural::src/core/runner/lab_args/tests.rs::HighItemCount",
         "structural::src/core/runner/session.rs::HighItemCount",
-        "structural::src/core/runner/worker.rs::GodFile",
         "structural::src/core/runner/workspace.rs::GodFile",
         "structural::src/core/runner/workspace.rs::HighItemCount",
         "structural::src/core/triage.rs::GodFile",
